requirements.txt
pytorch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
tqdm==4.65.0
numpy==1.26.0 
dgl-cuda11.8
pandas
pydantic
